requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.8
argcomplete==3.2.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
blinker==1.7.0
Brotli==1.0.9
cachetools==5.3.2
ccimport==0.4.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
chumpy==0.70
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.7
cumm==0.4.11
cumm-cu113==0.4.11
cv==1.0.0
cycler==0.12.1
Cython==3.0.8
dataclasses-json==0.6.3
decorator==5.1.1
Deprecated==1.2.14
diff-gaussian-rasterization==0.0.0
editdistance==0.6.2
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fairnr==0.0.0
fairseq==0.9.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.2
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
fvcore==0.1.5.post20221221
gmpy2==2.1.2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.20.2
humanfriendly==10.0
hydra-core==1.0.7
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
ipycanvas==0.13.1
ipyevents==2.0.2
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
kaolin==0.15.0
kiwisolver==1.4.5
kornia==0.7.1
lark==1.1.9
lazy_loader==0.3
lightning-utilities==0.10.0
llvmlite==0.41.1
lpips==0.1.4
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mesh-to-sdf==0.0.14
minimal-pytorch-rasterizer==0.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmcv-full==1.7.2
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.23.0
oauthlib==3.2.2
omegaconf==2.0.6
onnxruntime==1.16.3
open-clip-torch==2.0.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmesh==1.2.1
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pccm==0.4.11
pexpect==4.9.0
Pillow==10.0.1
pip==23.3.1
platformdirs==4.1.0
plyfile==1.0.3
pooch==1.8.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pyDeprecate==0.3.1
pyglet==2.0.12
pygltflib==1.16.1
Pygments==2.17.2
PyMatting==1.1.12
PyOpenGL==3.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrender==0.1.45
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.0
pytorch3d==0.7.5
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.32.1
regex==2023.12.25
rembg==2.0.53
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.17.1
rsa==4.9
sacrebleu==2.4.0
safetensors==0.4.1
scikit-image==0.16.2
scikit-learn==1.3.2
scipy==1.11.4
setuptools==60.2.0
six==1.16.0
smplx==0.1.28
spconv==2.3.5
spconv-cu113==2.3.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.3.0
torchvision==0.15.2
tornado==6.4
tqdm==4.65.2
traitlets==5.14.1
transformers==4.36.2
trimesh==4.0.9
triton==2.0.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
urllib3==1.26.18
usd-core==23.5
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==4.0.9
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
